feat: moved en tests to en folder and seperated the tests
