+++ firnas_vm/tests/en/mod.rs
@@ -1,0 +1,5 @@
+pub mod class_tests;
+pub mod control_flow_tests;
+pub mod function_tests;
+pub mod list_tests;
+pub mod simple_tests;
